+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * add $.terminal.parse_options which return same object as yargs parser
 * $.jrpc helper now return its own created promise instead of $.ajax
 * add wcwidth as dependency so it will always show wider characters correctly (in browsers will work the same as optional)
+* expose terminal exception in $.terminal namespace
 
 ### Bugfix
 * disable history in read & login (regression from 1.16.0 history interpreter option)